Refactor and cleanup com & native code
