+++ src/modules/peek/Peek.Common/Models/FileTime.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.InteropServices;
+
+namespace Peek.Common.Models
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct FileTime
+    {
+        public int DWHighDateTime;
+        public int DWLowDateTime;
+    }
+}
