+++ src/modules/peek/Peek.Common/Models/IEnumIDList.cs
@@ -1,0 +1,27 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Runtime.InteropServices;
+
+namespace Peek.Common.Models
+{
+    [ComImport]
+    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    [Guid("000214F2-0000-0000-C000-000000000046")]
+    public interface IEnumIDList
+    {
+        [PreserveSig]
+        int Next(int celt, out IntPtr rgelt, out int pceltFetched);
+
+        [PreserveSig]
+        int Skip(int celt);
+
+        [PreserveSig]
+        int Reset();
+
+        [PreserveSig]
+        int Clone(out IEnumIDList ppenum);
+    }
+}
