+++ src/modules/peek/Peek.Common/Models/IPropertyStore.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Runtime.InteropServices;
+using static Peek.Common.Helpers.PropertyStoreHelper;
+
+namespace Peek.Common.Models
+{
+    [ComImport]
+    [Guid("886D8EEB-8CF2-4446-8D02-CDBA1DBDCF99")]
+    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    public interface IPropertyStore
+    {
+        void GetCount(out uint propertyCount);
+
+        void GetAt(uint iProp, out PropertyKey pkey);
+
+        void GetValue(ref PropertyKey key, out PropVariant pv);
+
+        void SetValue(ref PropertyKey key, ref PropVariant pv);
+
+        void Commit();
+    }
+}
