+++ src/modules/peek/Peek.Common/Models/IShellItem.cs
@@ -4,7 +4,7 @@
 
 using System;
 using System.Runtime.InteropServices;
-using static Peek.Common.Models.PropertyStoreShellApi;
+using Windows.Win32.UI.Shell;
 
 namespace Peek.Common.Models
 {
@@ -27,17 +27,5 @@ namespace Peek.Common.Models
         void GetAttributes(uint sfgaoMask, out uint psfgaoAttribs);
 
         void Compare(IShellItem psi, uint hint, out int piOrder);
-    }
-
-    public enum Sigdn : uint
-    {
-        NormalDisplay = 0,
-        ParentRelativeParsing = 0x80018001,
-        ParentRelativeForAddressBar = 0x8001c001,
-        DesktopAbsoluteParsing = 0x80028000,
-        ParentRelativeEditing = 0x80031001,
-        DesktopAbsoluteEditing = 0x8004c000,
-        FileSysPath = 0x80058000,
-        Url = 0x80068000,
     }
 }
