+++ src/modules/peek/Peek.Common/NativeMethods.json
@@ -1,0 +1,4 @@
+﻿{
+    "$schema": "https://aka.ms/CsWin32.schema.json",
+    "public": true
+}
