+++ src/modules/peek/Peek.Common/NativeMethods.txt
@@ -1,0 +1,5 @@
+﻿GETPROPERTYSTOREFLAGS
+_SHCONTF
+SIGDN
+SHGDNF
+SIATTRIBFLAGS
