+++ src/modules/peek/Peek.UI/NativeMethods.txt
@@ -12,4 +12,3 @@ FindWindowEx
 SID_STopLevelBrowser
 GetClassName
 _SVGIO
-SIATTRIBFLAGS
